+++ MekHQ/src/mekhq/campaign/io/CampaignXmlParser.java
@@ -37,8 +37,8 @@ import mekhq.campaign.finances.Finances;
 import mekhq.campaign.force.Force;
 import mekhq.campaign.force.Lance;
 import mekhq.campaign.icons.UnitIcon;
-import mekhq.campaign.market.AbstractContractMarket;
-import mekhq.campaign.market.AtbMonthlyContractMarket;
+import mekhq.campaign.market.contractMarket.AbstractContractMarket;
+import mekhq.campaign.market.contractMarket.AtbMonthlyContractMarket;
 import mekhq.campaign.market.PersonnelMarket;
 import mekhq.campaign.market.ShoppingList;
 import mekhq.campaign.mission.AtBContract;
