+++ MekHQ/src/mekhq/campaign/market/contractMarket/DisabledContractMarket.java
@@ -1,9 +1,13 @@
-package mekhq.campaign.market;
+package mekhq.campaign.market.contractMarket;
 
 import mekhq.campaign.Campaign;
+import mekhq.campaign.market.enums.ContractMarketMethod;
 import mekhq.campaign.mission.AtBContract;
 
 public class DisabledContractMarket extends AbstractContractMarket {
+    public DisabledContractMarket() {
+        super(ContractMarketMethod.NONE);
+    }
 
     @Override
     public AtBContract addAtBContract(Campaign campaign) {
